[PM-11] Fix

+++ MPAPI/Sources/miet_practice_api/Migrations/CreateLookupTables.swift
@@ -3,32 +3,32 @@ import Fluent
 struct CreateLookupTables: AsyncMigration {
     func prepare(on database: any Database) async throws {
         try await database.schema(DatabaseSchema.userRole.rawValue)
-            .field("code", .int, .identifier(auto: false))
+            .field("code", .int8, .identifier(auto: false))
             .field("value", .string, .required)
             .create()
 
         try await database.schema(DatabaseSchema.testFramework.rawValue)
-            .field("code", .int, .identifier(auto: false))
+            .field("code", .int8, .identifier(auto: false))
             .field("value", .string, .required)
             .create()
 
         try await database.schema(DatabaseSchema.testResultStatus.rawValue)
-            .field("code", .int, .identifier(auto: false))
+            .field("code", .int8, .identifier(auto: false))
             .field("value", .string, .required)
             .create()
 
         try await database.schema(DatabaseSchema.taskStatus.rawValue)
-            .field("code", .int, .identifier(auto: false))
+            .field("code", .int8, .identifier(auto: false))
             .field("value", .string, .required)
             .create()
 
         try await database.schema(DatabaseSchema.notificationDeliveryStatus.rawValue)
-            .field("code", .int, .identifier(auto: false))
+            .field("code", .int8, .identifier(auto: false))
             .field("value", .string, .required)
             .create()
 
         try await database.schema(DatabaseSchema.buildStatus.rawValue)
-            .field("code", .int, .identifier(auto: false))
+            .field("code", .int8, .identifier(auto: false))
             .field("value", .string, .required)
             .create()
     }
